Update and support deletion on the engine class
