+++ project/include/graphics/GPUResource.h
@@ -14,6 +14,7 @@ namespace lime { namespace spoopy {
             operator const T &() const { return handle; }
             const T &GetHandle() const { return handle; }
 
+            virtual void Create() {}
             virtual void Destroy() {}
         protected:
             Device &device;
