+++ project/src/SpoopyExternalInterface.cpp
@@ -11,7 +11,6 @@
 
 #include "graphics/vulkan/GraphicsVulkan.h"
 #include "graphics/vulkan/ContextStage.h"
-#include "graphics/vulkan/RenderPassVulkan.h"
 #include "graphics/vulkan/components/CommandPoolVulkan.h"
 #include "graphics/vulkan/CommandBufferVulkan.h"
 #include "graphics/vulkan/QueueVulkan.h"
@@ -19,7 +18,6 @@
 
 
 typedef lime::spoopy::GraphicsVulkan GraphicsModule;
-typedef lime::spoopy::RenderPassVulkan RenderPass;
 typedef lime::spoopy::CommandPoolVulkan CommandPool;
 typedef lime::spoopy::CommandBufferVulkan CommandBuffer;
 typedef std::shared_ptr<lime::spoopy::QueueVulkan> Queue;
@@ -136,65 +134,6 @@ namespace lime { namespace spoopy {
     }
     DEFINE_PRIME3v(spoopy_device_set_swapchain_size);
 
-    void spoopy_add_color_attachment(value renderpass, int location, int format, bool hasImageLayout, bool sampled) {
-        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
-
-        #ifdef SPOOPY_VULKAN
-        VkSampleCountFlagBits samples = (GraphicsModule::GetCurrent()->MultisamplingEnabled && sampled)
-        ? GraphicsModule::GetCurrent()->GetPhysicalDevice()->GetMaxUsableSampleCount()
-        : VK_SAMPLE_COUNT_1_BIT;
-        VkImageLayout layout = hasImageLayout ? VK_IMAGE_LAYOUT_PRESENT_SRC_KHR : VK_IMAGE_LAYOUT_GENERAL;
-
-        _renderPass->AddColorAttachment(location, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL, format, samples, layout);
-        #endif
-    }
-    DEFINE_PRIME5v(spoopy_add_color_attachment);
-
-    void spoopy_add_depth_attachment(value renderpass, int location, int format, bool hasStencil, bool sampled) {
-        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
-
-        #ifdef SPOOPY_VULKAN
-        VkSampleCountFlagBits samples = (GraphicsModule::GetCurrent()->MultisamplingEnabled && sampled)
-        ? GraphicsModule::GetCurrent()->GetPhysicalDevice()->GetMaxUsableSampleCount()
-        : VK_SAMPLE_COUNT_1_BIT;
-
-        _renderPass->AddDepthAttachment(location, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, format, samples,
-            VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, VK_IMAGE_LAYOUT_UNDEFINED, hasStencil);
-        #endif
-    }
-    DEFINE_PRIME5v(spoopy_add_depth_attachment);
-
-    void spoopy_add_subpass_dependency(value renderpass, bool has_external1, bool has_external2, int srcStageMask,
-        int dstStageMask, int srcAccessMask, int dstAccessMask, int dependencyFlags) {
-        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
-
-        #ifdef SPOOPY_VULKAN
-        uint32_t _srcSubpass = has_external1 ? VK_SUBPASS_EXTERNAL : 0;
-        uint32_t _dstSubpass = has_external2 ? VK_SUBPASS_EXTERNAL : 0;
-        VkPipelineStageFlags _srcStageMask = (VkPipelineStageFlags)srcStageMask;
-        VkPipelineStageFlags _dstStageMask = (VkPipelineStageFlags)dstStageMask;
-        VkAccessFlags _srcAccessMask = (VkAccessFlags)srcAccessMask;
-        VkAccessFlags _dstAccessMask = (VkAccessFlags)dstAccessMask;
-        VkDependencyFlags _dependencyFlags = (VkDependencyFlags)dependencyFlags;
-
-        _renderPass->AddSubpassDependency(_srcSubpass, _dstSubpass, _srcStageMask, _dstStageMask, _srcAccessMask,
-            _dstAccessMask, _dependencyFlags);
-        #endif
-    }
-    DEFINE_PRIME8v(spoopy_add_subpass_dependency);
-
-    void spoopy_create_subpass(value renderpass) {
-        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
-        _renderPass->CreateSubpass();
-    }
-    DEFINE_PRIME1v(spoopy_create_subpass);
-
-    void spoopy_create_renderpass(value renderpass) {
-        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
-        _renderPass->CreateRenderPass();
-    }
-    DEFINE_PRIME1v(spoopy_create_renderpass);
-
     void spoopy_check_context(value window_handle) {
         Window* window = (Window*)val_data(window_handle);
         SDLWindow* sdlWindow = static_cast<SDLWindow*>(window);
@@ -220,11 +159,6 @@ namespace lime { namespace spoopy {
 
     // Objects
 
-    void spoopy_gc_render_pass(value handle) {
-        RenderPass* _renderPass = (RenderPass*)val_data(handle);
-        delete _renderPass;
-    }
-
     void spoopy_gc_command_pool(value handle) {
         CommandPool* _commandPool = (CommandPool*)val_data(handle);
         delete _commandPool;
@@ -244,12 +178,6 @@ namespace lime { namespace spoopy {
         Entry* _entry = (Entry*)val_data(handle);
         delete _entry;
     }
-
-    value spoopy_create_render_pass() {
-        RenderPass* _renderPass = new RenderPass(*GraphicsModule::GetCurrent()->GetLogicalDevice());
-        return CFFIPointer(_renderPass, spoopy_gc_render_pass);
-    }
-    DEFINE_PRIME0(spoopy_create_render_pass);
 
     /*
      * This is a interesting way of doing things,
