+++ project/src/cffi/UnsafeBindings.cpp
@@ -18,6 +18,7 @@
 #include "../graphics/vulkan/GraphicsVulkan.h"
 #include "../graphics/vulkan/components/SemaphoreVulkan.h"
 #include "../graphics/vulkan/PipelineVulkan.h"
+#include "../graphics/vulkan/RenderPassVulkan.h"
 #include "../graphics/vulkan/components/FenceVulkan.h"
 #include "../graphics/vulkan/CommandBufferVulkan.h"
 #include "../graphics/vulkan/QueueVulkan.h"
@@ -27,6 +28,7 @@ typedef lime::spoopy::GraphicsVulkan GraphicsModule;
 typedef lime::spoopy::PipelineVulkan Pipeline;
 typedef lime::spoopy::FenceVulkan GPUFence;
 typedef lime::spoopy::CommandBufferVulkan GPUCommandBuffer;
+typedef lime::spoopy::RenderPassVulkan RenderPass;
 
 #endif
 
@@ -47,6 +49,11 @@ namespace lime { namespace spoopy {
     void spoopy_delete_gpu_fence(value fence) {
         GPUFence* _fence = (GPUFence*)val_data(fence);
         delete _fence;
+    }
+
+    void spoopy_gc_render_pass(value handle) {
+        RenderPass* _renderPass = (RenderPass*)val_data(handle);
+        delete _renderPass;
     }
 
 
@@ -73,6 +80,13 @@ namespace lime { namespace spoopy {
         return CFFIPointer(_fence, spoopy_delete_gpu_fence);
     }
     DEFINE_PRIME1(spoopy_create_gpu_fence);
+
+    value spoopy_create_render_pass() {
+        RenderPass* _renderPass = new RenderPass(*GraphicsModule::GetCurrent()->GetLogicalDevice());
+        return CFFIPointer(_renderPass, spoopy_gc_render_pass);
+    }
+    DEFINE_PRIME0(spoopy_create_render_pass);
+
 
     void spoopy_set_gpu_fence_signal(value fence, bool signaled) {
         GPUFence* _fence = (GPUFence*)val_data(fence);
@@ -172,6 +186,71 @@ namespace lime { namespace spoopy {
         );
     }
     DEFINE_PRIME5(spoopy_queue_submit);
+
+    void spoopy_add_color_attachment(value renderpass, int location, int format, bool hasImageLayout, bool sampled) {
+    RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
+
+        #ifdef SPOOPY_VULKAN
+
+        VkSampleCountFlagBits samples = (GraphicsModule::GetCurrent()->MultisamplingEnabled && sampled)
+        ? GraphicsModule::GetCurrent()->GetPhysicalDevice()->GetMaxUsableSampleCount()
+        : VK_SAMPLE_COUNT_1_BIT;
+        VkImageLayout layout = hasImageLayout ? VK_IMAGE_LAYOUT_PRESENT_SRC_KHR : VK_IMAGE_LAYOUT_GENERAL;
+
+        _renderPass->AddColorAttachment(location, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL, format, samples, layout);
+
+        #endif
+    }
+    DEFINE_PRIME5v(spoopy_add_color_attachment);
+
+    void spoopy_add_depth_attachment(value renderpass, int location, int format, bool hasStencil, bool sampled) {
+    RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
+
+        #ifdef SPOOPY_VULKAN
+
+        VkSampleCountFlagBits samples = (GraphicsModule::GetCurrent()->MultisamplingEnabled && sampled)
+        ? GraphicsModule::GetCurrent()->GetPhysicalDevice()->GetMaxUsableSampleCount()
+        : VK_SAMPLE_COUNT_1_BIT;
+
+        _renderPass->AddDepthAttachment(location, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, format, samples,
+            VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, VK_IMAGE_LAYOUT_UNDEFINED, hasStencil);
+
+        #endif
+    }
+    DEFINE_PRIME5v(spoopy_add_depth_attachment);
+
+    void spoopy_add_subpass_dependency(value renderpass, bool has_external1, bool has_external2, int srcStageMask,
+    int dstStageMask, int srcAccessMask, int dstAccessMask, int dependencyFlags) {
+        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
+
+        #ifdef SPOOPY_VULKAN
+
+        uint32_t _srcSubpass = has_external1 ? VK_SUBPASS_EXTERNAL : 0;
+        uint32_t _dstSubpass = has_external2 ? VK_SUBPASS_EXTERNAL : 0;
+        VkPipelineStageFlags _srcStageMask = (VkPipelineStageFlags)srcStageMask;
+        VkPipelineStageFlags _dstStageMask = (VkPipelineStageFlags)dstStageMask;
+        VkAccessFlags _srcAccessMask = (VkAccessFlags)srcAccessMask;
+        VkAccessFlags _dstAccessMask = (VkAccessFlags)dstAccessMask;
+        VkDependencyFlags _dependencyFlags = (VkDependencyFlags)dependencyFlags;
+
+        _renderPass->AddSubpassDependency(_srcSubpass, _dstSubpass, _srcStageMask, _dstStageMask, _srcAccessMask,
+            _dstAccessMask, _dependencyFlags);
+
+        #endif
+    }
+    DEFINE_PRIME8v(spoopy_add_subpass_dependency);
+
+    void spoopy_create_subpass(value renderpass) {
+        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
+        _renderPass->CreateSubpass();
+    }
+    DEFINE_PRIME1v(spoopy_create_subpass);
+
+    void spoopy_create_renderpass(value renderpass) {
+        RenderPass* _renderPass = (RenderPass*)val_data(renderpass);
+        _renderPass->Create();
+    }
+    DEFINE_PRIME1v(spoopy_create_renderpass);
 
     void spoopy_dealloc_gpu_cffi_pointer(int type, value pointer) { // Pure Deallocator for raw references
         switch((BackendType)type) {
